Restructured InventoryManager
