+++ project.godot
@@ -12,7 +12,7 @@ config_version=5
 
 config/name="NewGame"
 run/main_scene="res://scenes/world.tscn"
-config/features=PackedStringArray("4.2", "C#", "GL Compatibility")
+config/features=PackedStringArray("4.1", "C#", "GL Compatibility")
 config/icon="res://icon.svg"
 
 [autoload]
