+++ scenes/world.tscn
@@ -1620,10 +1620,25 @@ format = 2
 layer_0/name = "Bg"
 layer_0/tile_data = PackedInt32Array(0, 524288, 9, 65536, 524288, 10, 131072, 524288, 11, 1, 589824, 9, 65537, 589824, 10, 131073, 589824, 11, 2, 655360, 9, 65538, 655360, 10, 131074, 655360, 11, 3, 589824, 9, 65539, 589824, 10, 131075, 589824, 11, 4, 589824, 9, 65540, 589824, 10, 131076, 589824, 11, 5, 589824, 9, 65541, 589824, 10, 131077, 589824, 11, 6, 655360, 9, 65542, 655360, 10, 131078, 655360, 11, 7, 589824, 9, 65543, 589824, 10, 131079, 589824, 11, 8, 589824, 9, 131080, 589824, 11, 9, 589824, 9, 65545, 589824, 10, 131081, 589824, 11, 10, 589824, 9, 65546, 589824, 10, 131082, 589824, 11, 11, 589824, 9, 131083, 589824, 11, 12, 589824, 9, 65548, 589824, 10, 131084, 589824, 11, 13, 589824, 9, 65549, 589824, 10, 131085, 589824, 11, 196608, 327680, 17, 196621, 65536, 20, 196620, 65536, 20, 196619, 65536, 20, 262155, 65536, 21, 262154, 0, 21, 262153, 393216, 18, 262152, 393216, 18, 196616, 393216, 17, 196615, 393216, 17, 196617, 393216, 17, 196618, 0, 20, 196614, 393216, 17, 196613, 393216, 17, 196612, 393216, 17, 196611, 393216, 17, 196610, 393216, 17, 196609, 393216, 17, 262144, 327680, 18, 327680, 327680, 18, 393216, 327680, 18, 458752, 327680, 18, 458765, 393216, 18, 262157, 65536, 21, 327693, 65536, 21, 393229, 65536, 22, 262156, 65536, 21, 262151, 393216, 18, 262150, 393216, 18, 262149, 393216, 18, 262148, 393216, 18, 262147, 393216, 18, 262146, 393216, 18, 262145, 393216, 18, 327681, 393216, 18, 327682, 393216, 18, 327683, 393216, 18, 327684, 393216, 18, 327685, 393216, 18, 327686, 393216, 18, 327687, 393216, 18, 327688, 393216, 18, 327689, 393216, 18, 327690, 0, 21, 327691, 65536, 21, 327692, 65536, 21, 393228, 65536, 22, 393227, 65536, 22, 393226, 0, 22, 393225, 393216, 18, 393224, 393216, 18, 393223, 393216, 18, 393222, 393216, 18, 393221, 393216, 18, 393220, 393216, 18, 393219, 393216, 18, 393218, 393216, 18, 393217, 393216, 18, 458764, 393216, 18, 458763, 393216, 18, 458762, 393216, 18, 458761, 393216, 18, 458760, 393216, 18, 458759, 393216, 18, 458758, 393216, 18, 458757, 393216, 18, 458756, 393216, 18, 458755, 393216, 18, 458754, 393216, 18, 458753, 393216, 18, 14, 589824, 9, 65550, 589824, 10, 131086, 589824, 11, 15, 589824, 9, 65551, 589824, 10, 131087, 589824, 11, 16, 589824, 9, 65552, 589824, 10, 131088, 589824, 11, 17, 720896, 9, 65553, 720896, 10, 131089, 720896, 11, 196625, 458752, 17, 262161, 458752, 18, 327697, 458752, 18, 393233, 458752, 18, 458769, 458752, 18, 524305, 458752, 18, 589841, 458752, 19, 196622, 131072, 20, 196623, 393216, 17, 196624, 393216, 17, 589840, 393216, 19, 589839, 393216, 19, 589838, 393216, 19, 589837, 393216, 19, 589836, 393216, 19, 589835, 393216, 19, 589834, 393216, 19, 589833, 393216, 19, 589832, 393216, 19, 589831, 393216, 19, 589830, 393216, 19, 589829, 393216, 19, 589828, 393216, 19, 589827, 393216, 19, 589826, 393216, 19, 589825, 393216, 19, 589824, 327680, 19, 524288, 327680, 18, 524289, 393216, 18, 524290, 393216, 18, 524291, 393216, 18, 524292, 393216, 18, 524293, 393216, 18, 524294, 393216, 18, 524295, 393216, 18, 524296, 393216, 18, 524297, 393216, 18, 524298, 393216, 18, 524299, 393216, 18, 524300, 393216, 18, 524301, 393216, 18, 524302, 393216, 18, 524303, 393216, 18, 524304, 393216, 18, 458768, 393216, 18, 458767, 393216, 18, 393231, 393216, 18, 393230, 131072, 22, 458766, 393216, 18, 262158, 131072, 21, 262159, 393216, 18, 262160, 393216, 18, 327696, 393216, 18, 327695, 393216, 18, 327694, 131072, 21, 393232, 393216, 18, 65547, 589824, 10, 65544, 589824, 10)
 layer_1/name = "chair"
+layer_1/enabled = true
+layer_1/modulate = Color(1, 1, 1, 1)
+layer_1/y_sort_enabled = false
+layer_1/y_sort_origin = 0
+layer_1/z_index = 0
 layer_1/tile_data = PackedInt32Array(327684, 131072, 54, 393220, 131072, 55, 327682, 65536, 54, 393218, 65536, 55)
 layer_2/name = "Furniture"
+layer_2/enabled = true
+layer_2/modulate = Color(1, 1, 1, 1)
+layer_2/y_sort_enabled = false
+layer_2/y_sort_origin = 0
+layer_2/z_index = 0
 layer_2/tile_data = PackedInt32Array(131075, 393216, 52, 196611, 393216, 53, 131076, 458752, 52, 196612, 458752, 53, 327682, 0, 43, 393218, 0, 44, 458754, 0, 45, 327683, 65536, 43, 393219, 65536, 44, 458755, 65536, 45, 327684, 131072, 43, 393220, 131072, 44, 458756, 131072, 45, 65537, 393216, 49, 131073, 393216, 50, 196609, 393216, 51, 65538, 458752, 49, 131074, 458752, 50, 196610, 458752, 51, 65541, 196608, 40, 131072, 393216, 56, 196608, 393216, 57, 65548, 524288, 42, 131084, 524288, 43, 196620, 524288, 44, 65549, 589824, 42, 131085, 589824, 43, 196621, 589824, 44, 131083, 393216, 45, 196619, 393216, 46, 15, 131072, 12, 65551, 131072, 13, 9, 131072, 12, 65545, 131072, 13)
 layer_3/name = "Deco"
+layer_3/enabled = true
+layer_3/modulate = Color(1, 1, 1, 1)
+layer_3/y_sort_enabled = false
+layer_3/y_sort_origin = 0
+layer_3/z_index = 0
 layer_3/tile_data = PackedInt32Array(393219, 0, 49, 65537, 65536, 50, 131083, 196608, 43, 15, 720896, 0, 65551, 720896, 1, 131087, 720896, 2, 196623, 720896, 3, 9, 720896, 0, 65545, 720896, 1, 131081, 720896, 2, 196617, 720896, 3)
 
 [node name="Player" parent="." instance=ExtResource("1_c1fn1")]
